Disable latitude to meters / degree longitude test.
The test tests against known correct values that use WGS84 world model,
where as the code uses a spherical model. The code is probably close
enough for the needs of the AVC; it's off by several hundred meters over
a distance of 80 km, which I don't think the RC car is going to be
driving.