Start working on waypoint generator.
Compilation still fails because I'm not sure how to parse strings into
floats in Rust 0.12, which is the version I have on the Pi. From what I
can tell, the parsing method was changed between 0.12 and 1.0-alpha. I'm
not sure what the best plan of action is now, because I really don't
want to cross compile to the Pi but I'm not too crazy about doing test
development on my laptop with an older version of Rust, because part of
my goal was to learn Rust. Also, the build or Rust I'm using on the Pi
doesn't include Cargo. I think I might abandon the Rust port until
Debian or somebody else releases a Raspberry Pi port, which will
probably be when 1.0 is released in a few months.